pyproject.toml
[project]
name = "vc-data-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.0.0",
    "langchain-mcp-adapters>=0.1.0",
    "langgraph>=0.2.0",
    "langchain-groq>=0.1.0",
]

[project.scripts]
vc-data-mcp = "server:main"
